cluster/streamlit_docker/requirements.txt
pandas
dataiku-api-client
streamlit